Move ginkgo setup suite to own file.
The other e2e suites do this already.
