+++ clients/gitlabrepo/graphql_test.go
@@ -17,22 +17,11 @@ package gitlabrepo
 import (
 	"context"
 	"fmt"
-	"os"
-	"testing"
 	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-func TestGitlabRepoE2E(t *testing.T) {
-	if val, exists := os.LookupEnv("SKIP_GINKGO"); exists && val == "1" {
-		t.Skip()
-	}
-	t.Parallel()
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "GitLab Repo Suite")
-}
 
 var _ = Describe("E2E TEST: gitlabrepo.graphqlHandler", func() {
 	var graphqlhandler graphqlHandler
